# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cloud-api/tenants": {
      "post": {
        "operationId": "TenantController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Tenant creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation failed"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          },
          "409": {
            "description": "Conflict - Tenant with subdomain already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new tenant",
        "tags": [
          "Tenants"
        ]
      },
      "get": {
        "operationId": "TenantController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all tenants retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve all tenants",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/cloud-api/tenants/{id}": {
      "get": {
        "operationId": "TenantController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the tenant",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a tenant by ID",
        "tags": [
          "Tenants"
        ]
      },
      "patch": {
        "operationId": "TenantController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the tenant to update",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Tenant update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation failed"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          },
          "404": {
            "description": "Tenant not found"
          },
          "409": {
            "description": "Conflict - Subdomain already in use by another tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a tenant",
        "tags": [
          "Tenants"
        ]
      },
      "delete": {
        "operationId": "TenantController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the tenant to archive",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant archived successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a tenant (soft delete)",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/cloud-api/tenants/subdomain/{subdomain}": {
      "get": {
        "operationId": "TenantController_findBySubdomain",
        "parameters": [
          {
            "name": "subdomain",
            "required": true,
            "in": "path",
            "description": "Unique subdomain identifier of the tenant",
            "schema": {
              "example": "acme-corp",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a tenant by subdomain",
        "tags": [
          "Tenants"
        ]
      }
    },
...
    "/cloud-api/onboarding/2fa/totp/setup": {
      "post": {
        "operationId": "TwoFactorController_initiateTotpSetup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns QR code and manual key for TOTP setup"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          },
          "409": {
            "description": "TOTP already configured for this user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate TOTP (Time-based One-Time Password) setup",
        "tags": [
          "Onboarding - Two-Factor Authentication"
        ]
      }
    },
    "/cloud-api/onboarding/2fa/totp/verify": {
      "post": {
        "operationId": "TwoFactorController_verifyTotpSetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "TOTP verification payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTotpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TOTP verified successfully, returns backup codes"
          },
          "400": {
            "description": "Invalid TOTP code"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify TOTP setup with a code from authenticator app",
        "tags": [
          "Onboarding - Two-Factor Authentication"
        ]
      }
    },
    "/cloud-api/onboarding/2fa/skip": {
      "post": {
        "operationId": "TwoFactorController_skip2FASetup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "2FA setup skipped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Two-factor authentication setup skipped. You can enable it later in settings."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Skip two-factor authentication setup",
        "tags": [
          "Onboarding - Two-Factor Authentication"
        ]
      }
    },
    "/cloud-api/onboarding/2fa/status": {
      "get": {
        "operationId": "TwoFactorController_get2FAStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current 2FA configuration status"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current two-factor authentication status",
        "tags": [
          "Onboarding - Two-Factor Authentication"
        ]
      }
    },
    "/cloud-api/onboarding/2fa/passkey/setup": {
      "post": {
        "operationId": "TwoFactorController_initiatePasskeySetup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns WebAuthn registration options for passkey setup"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          },
          "409": {
            "description": "Passkey already configured for this user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate Passkey/WebAuthn setup",
        "tags": [
          "Onboarding - Two-Factor Authentication"
        ]
      }
    },
    "/cloud-api/onboarding/2fa/passkey/verify": {
      "post": {
        "operationId": "TwoFactorController_verifyPasskeySetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Passkey credential verification payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPasskeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Passkey registered successfully, returns backup codes"
          },
          "400": {
            "description": "Invalid passkey credential"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing onboarding token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify Passkey/WebAuthn credential registration",
        "tags": [
          "Onboarding - Two-Factor Authentication"
        ]
      }
    }
  },
  "info": {
    "title": "Vritti Cloud API",
    "description": "Internal API for Vritti SaaS Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "subdomain": {
            "type": "string",
            "description": "Unique subdomain identifier for the tenant. Used in URLs and routing.",
            "example": "acme-corp",
            "minLength": 2,
            "maxLength": 50,
            "pattern": "^[a-z0-9-]+$"
          },
          "name": {
            "type": "string",
            "description": "Display name of the tenant organization",
            "example": "Acme Corporation",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Optional description of the tenant organization",
            "example": "Leading provider of innovative solutions",
            "maxLength": 500
          },
          "dbType": {
            "type": "string",
            "description": "Database deployment type for the tenant",
            "enum": [
              "SHARED",
              "DEDICATED"
            ],
            "example": "SHARED"
          },
          "status": {
            "type": "string",
            "description": "Current operational status of the tenant",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "ARCHIVED"
            ],
            "example": "ACTIVE",
            "default": "ACTIVE"
          },
          "dbHost": {
            "type": "string",
            "description": "Database server hostname or IP address (required for DEDICATED tenants)",
            "example": "db.acme-corp.example.com"
          },
          "dbPort": {
            "type": "number",
            "description": "Database server port number",
            "example": 5432,
            "minimum": 1,
            "maximum": 65535
          },
          "dbUsername": {
            "type": "string",
            "description": "Database username for authentication",
            "example": "tenant_user"
          },
          "dbPassword": {
            "type": "string",
            "description": "Database password for authentication",
            "example": "secure_password_123"
          },
          "dbName": {
            "type": "string",
            "description": "Name of the database to connect to",
            "example": "acme_corp_db"
          },
          "dbSchema": {
            "type": "string",
            "description": "Database schema name (for databases that support schemas)",
            "example": "tenant_schema"
          },
          "dbSslMode": {
            "type": "string",
            "description": "SSL mode for database connection security",
            "enum": [
              "require",
              "prefer",
              "disable"
            ],
            "example": "require"
          },
          "connectionPoolSize": {
            "type": "number",
            "description": "Maximum number of connections in the database connection pool",
            "example": 10,
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "subdomain",
          "name",
          "dbType"
        ]
      },
      "TenantDatabaseConfigResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the database configuration",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "tenantId": {
            "type": "string",
            "description": "Reference to the parent tenant",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dbHost": {
            "type": "string",
            "description": "Database server hostname or IP address",
            "example": "db.acme-corp.example.com"
          },
          "dbPort": {
            "type": "number",
            "description": "Database server port number",
            "example": 5432,
            "minimum": 1,
            "maximum": 65535
          },
          "dbName": {
            "type": "string",
            "description": "Name of the database",
            "example": "acme_corp_db"
          },
          "dbSchema": {
            "type": "object",
            "description": "Database schema name (for databases that support schemas)",
            "example": "tenant_schema",
            "nullable": true
          },
          "dbSslMode": {
            "type": "string",
            "description": "SSL mode for database connection security",
            "enum": [
              "require",
              "prefer",
              "disable"
            ],
            "example": "require"
          },
          "connectionPoolSize": {
            "type": "number",
            "description": "Maximum number of connections in the database connection pool",
            "example": 10,
            "minimum": 1,
            "maximum": 100
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when the configuration was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when the configuration was last updated",
            "example": "2024-01-20T14:45:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "dbHost",
          "dbPort",
          "dbName",
          "dbSslMode",
          "connectionPoolSize",
          "createdAt",
          "updatedAt"
        ]
      },
      "TenantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the tenant",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "subdomain": {
            "type": "string",
            "description": "Unique subdomain identifier for the tenant. Used in URLs and routing.",
            "example": "acme-corp"
          },
          "name": {
            "type": "string",
            "description": "Display name of the tenant organization",
            "example": "Acme Corporation"
          },
          "description": {
            "type": "object",
            "description": "Optional description of the tenant organization",
            "example": "Leading provider of innovative solutions",
            "nullable": true
          },
          "dbType": {
            "type": "string",
            "description": "Database deployment type for the tenant",
            "enum": [
              "SHARED",
              "DEDICATED"
            ],
            "example": "SHARED"
          },
          "status": {
            "type": "string",
            "description": "Current operational status of the tenant",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "ARCHIVED"
            ],
            "example": "ACTIVE"
          },
          "databaseConfig": {
            "description": "Database configuration details (only present for DEDICATED tenants)",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TenantDatabaseConfigResponseDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when the tenant was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when the tenant was last updated",
            "example": "2024-01-20T14:45:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "subdomain",
          "name",
          "dbType",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "subdomain": {
            "type": "string",
            "description": "Unique subdomain identifier for the tenant. Used in URLs and routing.",
            "example": "acme-corp",
            "minLength": 2,
            "maxLength": 50,
            "pattern": "^[a-z0-9-]+$"
          },
          "name": {
            "type": "string",
            "description": "Display name of the tenant organization",
            "example": "Acme Corporation",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Optional description of the tenant organization",
            "example": "Leading provider of innovative solutions",
            "maxLength": 500
          },
          "dbType": {
            "type": "string",
            "description": "Database deployment type for the tenant",
            "enum": [
              "SHARED",
              "DEDICATED"
            ],
            "example": "SHARED"
          },
          "status": {
            "type": "string",
            "description": "Current operational status of the tenant",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "ARCHIVED"
            ],
            "example": "ACTIVE",
            "default": "ACTIVE"
          },
          "dbHost": {
            "type": "string",
            "description": "Database server hostname or IP address (required for DEDICATED tenants)",
            "example": "db.acme-corp.example.com"
          },
          "dbPort": {
            "type": "number",
            "description": "Database server port number",
            "example": 5432,
            "minimum": 1,
            "maximum": 65535
          },
          "dbUsername": {
            "type": "string",
            "description": "Database username for authentication",
            "example": "tenant_user"
          },
          "dbPassword": {
            "type": "string",
            "description": "Database password for authentication",
            "example": "secure_password_123"
          },
          "dbName": {
            "type": "string",
            "description": "Name of the database to connect to",
            "example": "acme_corp_db"
          },
          "dbSchema": {
            "type": "string",
            "description": "Database schema name (for databases that support schemas)",
            "example": "tenant_schema"
          },
          "dbSslMode": {
            "type": "string",
            "description": "SSL mode for database connection security",
            "enum": [
              "require",
              "prefer",
              "disable"
            ],
            "example": "require"
          },
          "connectionPoolSize": {
            "type": "number",
            "description": "Maximum number of connections in the database connection pool",
            "example": 10,
            "minimum": 1,
            "maximum": 100
          }
        }
      },
...
      "VerifyTotpDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Six-digit TOTP verification code from the authenticator app",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "token"
        ]
      },
      "VerifyPasskeyDto": {
        "type": "object",
        "properties": {
          "credential": {
            "type": "object",
            "description": "WebAuthn credential registration response from the browser containing attestation data",
            "example": {
              "id": "base64url-credential-id",
              "rawId": "base64url-raw-credential-id",
              "response": {
                "clientDataJSON": "base64url-client-data",
                "attestationObject": "base64url-attestation-object",
                "transports": [
                  "internal",
                  "hybrid"
                ]
              },
              "authenticatorAttachment": "platform",
              "clientExtensionResults": {},
              "type": "public-key"
            }
          }
        },
        "required": [
          "credential"
        ]
      }
    }
  }
}